Seperate Drating from PlayerHome
